# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce API",
    "version": "1.0.0",
    "description": "API documentation for the E-Commerce platform"
  },
  "servers": [
    {
      "url": "http://localhost:3300"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/vouchers": {
      "post": {
        "summary": "Create a new voucher",
        "tags": [
          "Vouchers"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "discount",
                  "valid_from",
                  "valid_until"
                ],
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "discount": {
                    "type": "number"
                  },
                  "valid_from": {
                    "type": "string",
                    "format": "date"
                  },
                  "valid_until": {
                    "type": "string",
                    "format": "date"
                  },
                  "user_id": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all available vouchers",
        "tags": [
          "Vouchers"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of vouchers"
          }
        }
      }
    },
    "/api/vouchers/me": {
      "get": {
        "summary": "Get available vouchers member",
        "tags": [
          "Vouchers"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of vouchers"
          }
        }
      }
    },
    "/api/vouchers-publish": {
      "get": {
        "summary": "Get all available vouchers",
        "tags": [
          "Vouchers"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of vouchers"
          }
        }
      }
    },
    "/api/vouchers/{id}": {
      "delete": {
        "summary": "Soft delete a voucher",
        "tags": [
          "Vouchers"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Voucher deleted"
          }
        }
      },
      "put": {
        "summary": "Update a voucher",
        "tags": [
          "Vouchers"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "discount": {
                    "type": "number"
                  },
                  "valid_from": {
                    "type": "string",
                    "format": "date"
                  },
                  "valid_until": {
                    "type": "string",
                    "format": "date"
                  },
                  "user_id": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/vouchers/code/{code}": {
      "get": {
        "summary": "Get voucher by code",
        "tags": [
          "Vouchers"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Voucher found"
          }
        }
      }
    },
    "/api/voucher/validate": {
      "get": {
        "summary": "Validasi voucher berdasarkan kode dan (opsional) user ID",
        "tags": [
          "Voucher"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Kode voucher"
          },
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "required": false,
            "description": "ID user (jika voucher bersifat personal)"
          }
        ],
        "responses": {
          "200": {
            "description": "Validasi voucher berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean",
                      "example": true
                    },
                    "voucher": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "code": {
                          "type": "string",
                          "example": "VOUCHER50K"
                        },
                        "discount_amount": {
                          "type": "number",
                          "format": "float",
                          "example": 50000
                        },
                        "min_purchase": {
                          "type": "number",
                          "format": "float",
                          "example": 100000
                        },
                        "valid_from": {
                          "type": "string",
                          "format": "date",
                          "example": "2025-04-01"
                        },
                        "valid_until": {
                          "type": "string",
                          "format": "date",
                          "example": "2025-04-30"
                        },
                        "user_id": {
                          "type": "integer",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Voucher tidak valid atau tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Voucher tidak ditemukan, sudah digunakan, atau tidak berlaku untuk user ini"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Terjadi kesalahan internal saat validasi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan saat memvalidasi voucher"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}